test interfaces to avoid replications + more strict swagger definitions/validation
