+++ magpie/api/login/login.py
@@ -209,7 +209,7 @@ def authomatic_login(request):
     return response
 
 
-@SessionAPI.get(schema=Session_GET_ResponseBodySchema(), tags=[LoginTag], response_schemas={
+@SessionAPI.get(tags=[LoginTag], response_schemas={
     '200': Session_GET_OkResponseSchema(),
     '500': Session_GET_InternalServerErrorResponseSchema()
 })
